[core] Merge geojson_tile_data.{hpp,cpp} into geojson_tile.{hpp,cpp}
