+++ src/mbgl/tile/geojson_tile.hpp
@@ -1,8 +1,30 @@
 #pragma once
 
 #include <mbgl/tile/geometry_tile.hpp>
+#include <mbgl/tile/geometry_tile_data.hpp>
+
+namespace mapbox {
+namespace geojsonvt {
+class GeoJSONVT;
+} // namespace geojsonvt
+} // namespace mapbox
 
 namespace mbgl {
+
+namespace style {
+class UpdateParameters;
+}
+
+class GeoJSONTileData : public GeometryTileData {
+public:
+    GeoJSONTileData(const OverscaledTileID&,
+                   std::string sourceID,
+                   const style::UpdateParameters&,
+                   mapbox::geojsonvt::GeoJSONVT*);
+    ~GeoJSONTileData();
+
+    void setNecessity(Necessity) final;
+};
 
 // Implements a simple in-memory Tile type that holds GeoJSON values. A GeoJSON tile can only have
 // one layer, and it is always returned regardless of which layer is requested.
